+++ packages/nodejs/test/cluster/ClusterServerTestingUtil.ts
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { PrivateKey } from "#general";
 import { Message, SecureSession } from "#protocol";
 import { NodeId, StatusCode } from "#types";
 import { asClusterServerInternal, ClusterServerObj, ClusterType } from "@project-chip/matter.js/cluster";
@@ -12,9 +11,6 @@ import { Endpoint } from "@project-chip/matter.js/device";
 import { createTestFabric } from "../interaction/InteractionTestUtils.js";
 
 export const ZERO = new Uint8Array(1);
-const PRIVATE_KEY = new Uint8Array(32);
-PRIVATE_KEY[31] = 1; // EC doesn't like all-zero private key
-export const KEY = PrivateKey(PRIVATE_KEY);
 
 // TODO make that nicer
 export async function callCommandOnClusterServer<T extends ClusterType>(