+++ packages/tools/src/index.ts
@@ -4,12 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import "./testing/global-declarations.js";
+export * from "./building/builder.js";
 export * from "./building/graph.js";
 export * from "./building/project.js";
-export * from "./testing/chip.js";
-export * from "./testing/mocharc.cjs";
-export * from "./util/docker.js";
+export * from "./util/commander.js";
+export * from "./util/files.js";
 export * from "./util/package.js";
 export * from "./util/progress.js";
-export * from "./util/wtf.js";