+++ packages/tools/src/util/package.ts
@@ -6,10 +6,9 @@
 
 import { existsSync, readFileSync, statSync } from "fs";
 import { readdir, readFile, stat, writeFile } from "fs/promises";
-import { glob } from "glob";
 import { dirname, join, relative, resolve } from "path";
-import { maybeStatSync } from "../testing/files.js";
 import { ignoreError, ignoreErrorSync } from "./errors.js";
+import { globSync, maybeStatSync } from "./files.js";
 import { Progress } from "./progress.js";
 import { toolsPath } from "./tools-path.cjs";
 
@@ -89,19 +88,24 @@ export class Package {
         return this.hasDirectory(CODEGEN_PATH);
     }
 
-    resolve(path: string) {
-        return resolve(this.path, path);
+    resolve(...paths: string[]) {
+        return resolve(this.path, ...paths);
     }
 
     relative(path: string) {
         return relative(this.path, path);
     }
 
-    async glob(pattern: string) {
+    async glob(pattern: string | string[]) {
         // Glob only understands forward-slash as separator because reasons
-        pattern = this.resolve(pattern).replace(/\\/g, "/");
+        if (typeof pattern === "string") {
+            pattern = this.resolve(pattern).replace(/\\/g, "/");
+        } else {
+            pattern = pattern.map(s => this.resolve(s).replace(/\\/g, "/"));
+        }
 
-        return await glob(pattern);
+        // Current glob implementation isn't actually async as this is faster and we only walk small directory trees
+        return globSync(pattern);
     }
 
     start(what: string) {
@@ -203,26 +207,31 @@ export class Package {
         throw new Error(`Cannot resolve export ${name} in package ${this.name}`);
     }
 
-    resolveImport(name: string, type: "cjs" | "esm" = "esm") {
-        const segments = name.split("/");
-        let subdir = segments.shift() as string;
-        if (subdir.startsWith("@") && segments.length) {
-            subdir = `${subdir}/${segments.shift()}`;
-        }
-
+    findPackage(name: string) {
         let resolveIn = this.path;
         while (true) {
-            if (isDirectory(resolve(resolveIn, "node_modules", subdir))) {
+            if (isDirectory(resolve(resolveIn, "node_modules", name))) {
                 break;
             }
             const nextResolveIn = dirname(resolveIn);
             if (nextResolveIn === resolveIn) {
-                throw new Error(`Cannot find module ${subdir} from ${this.path}`);
+                throw new Error(`Cannot find module ${name} from ${this.path}`);
             }
             resolveIn = nextResolveIn;
         }
 
-        const pkg = Package.forPath(resolve(resolveIn, "node_modules", subdir));
+        return Package.forPath(resolve(resolveIn, "node_modules", name));
+    }
+
+    resolveImport(name: string, type: "cjs" | "esm" = "esm") {
+        const segments = name.split("/");
+        let packageName = segments.shift() as string;
+        if (packageName.startsWith("@") && segments.length) {
+            packageName = `${packageName}/${segments.shift()}`;
+        }
+
+        const pkg = this.findPackage(packageName);
+
         return pkg.resolveExport(segments.length ? segments.join("/") : ".", type);
     }
 
@@ -259,8 +268,8 @@ export class Package {
                 e = new Error(`${e}`);
             }
             (e as Error).message = `Error parsing "${this.resolve(path)}": ${(e as Error).message}`;
+            throw e;
         }
-        return JSON.parse(await this.readFile(path));
     }
 
     async writeJson(path: string, value: {}) {